fix: sign the entire batch of events when received along with user DID.
This fixes a performance problem where signing was taking a long
time when importing large batches of events, and it also fixes
a signature problem where we were only signing the payload, not
the user that we knew sent the event.

This isn't perfect, though, as we are storing the same signature for
a whole bunch of events, when maybe we could only store it once
for the whole batch, and we are not including the event index
in the signature, which we should.